requirements.txt
streamlit
google-generativeai
groq
python-dotenv
sqlalchemy
pandas
fpdf2
matplotlib
